Expanded equalsNotNull
It will now check both objects to see if the equal the other. Since we're not using generics to programmatically ensure that the two objects are of the same type, we should assume therefore that there's a possibility that the two objects *aren't* of the same type but *may* still work out to be equal. And so instead of just relying on the former's equals method (which may produce different results to the former's) we test both.