DepedencyGlue
De-abstract the enabled and disabled methods. Their purpose is to function as post enable setup and graceful pre-disable desconstruction, as it were, however some glues wont necessarily need to do this and so it's kind of rude to force them to have two empty overridden methods. I've also added a method which is meant to be used within glue functions to determine whether the glue is safe to use.