CoreConfigManager :: vastly improve
Class is no longer abstract, so plugins can use it as is if they don't want to implement their own parser, or create a useless and empty extension. I've also added a robust database parser that supports old and new keys. This doesn't replace the casting method, but rather offers a more robust and friendlier version that tells you when something is wrong, allows for optional values, and isn't so throw happy.